# openapi/v1/openapi.yaml
openapi: 3.0.3
info:
  title: Resume AI API
  description: |
    API for the Job Application Resume Customization Platform.
    Enables users to manage their professional profile, submit job descriptions,
    and generate customized, ATS-friendly resumes using AI.
  version: "1.0.0"
servers:
  - url: http://localhost:8000/api/v1
    description: Development server
  - url: https://api.resume-ai.com/api/v1
    description: Production server
...
paths:
...
  /resumes:
    get:
      tags: [Resumes]
      summary: List all resume generation requests for the user
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        '200':
          description: A list of resume generation requests.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResumeGenerationRequest'
        '401':
          $ref: '#/components/responses/AuthRequired'
    post:
      tags: [Resumes]
      summary: Trigger a new resume generation
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - name: Idempotency-Key
          in: header
          schema:
            type: string
            format: uuid
          description: Prevents duplicate generation requests.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [job_description_id, template_id]
              properties:
                job_description_id:
                  type: string
                  format: uuid
                template_id:
                  type: string
      responses:
        '202':
          description: Resume generation accepted and in progress.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResumeGenerationRequest'
        '400':
          $ref: '#/components/responses/InvalidPayload'
        '401':
          $ref: '#/components/responses/AuthRequired'
        '429':
          $ref: '#/components/responses/RateLimited'
...
components:
  schemas:
...
    ResumeGenerationRequest:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        status:
          type: string
          enum: [pending, processing, success, failed]
          readOnly: true
        failure_reason:
          type: string
          nullable: true
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        expires_at:
          type: string
          format: date-time
          readOnly: true
    Error:
      type: object
      required: [error_code, message]
      properties:
        error_code:
          type: string
          description: A machine-readable error code.
        message:
          type: string
          description: A human-readable error message.
        details:
          type: object
          additionalProperties: true
          nullable: true
...
  responses:
    AuthRequired:
      description: Authentication is required and has failed or has not yet been provided.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error_code: "AUTH_REQUIRED"
            message: "Authentication credentials were not provided."
    InvalidPayload:
      description: The request payload is invalid or missing required fields.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error_code: "INVALID_PAYLOAD"
            message: "Invalid input."
            details: { "email": ["Enter a valid email address."] }
...
    RateLimited:
      description: Too many requests.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error_code: "RATE_LIMITED"
            message: "Request was throttled. Try again later."
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT